fixing deps
