+++ lcli/src/http_sync.rs
@@ -1,6 +1,7 @@
 use clap::ArgMatches;
 use clap_utils::{parse_optional, parse_required};
 use environment::Environment;
+use eth2::Error;
 use eth2::{
     BeaconNodeHttpClient, SensitiveUrl, Timeouts,
     types::{BlockId, ChainSpec, ForkName, PublishBlockRequest, SignedBlockContents},
